refactor: update to cargo workspace + set cloudlet project name
