+++ Cargo.toml
@@ -1,0 +1,3 @@
+[workspace]
+members = ["src/vmm"]
+resolver = "2"
