+++ src/vmm/Cargo.toml
@@ -1,21 +1,25 @@
 [package]
-name = "harkness-vmm"
-description = "The VMM implementation for the Harkness serverless runtime"
+name = "vmm"
+description = "The Virtual Machine Manager for the Cloudlet serverless runtime"
 authors = ["Polytech Montpellier DevOps - 2022/2025"]
 license = "Apache-2.0"
 readme = "README.md"
-repository = "https://github.com/virt-do/project-2024"
+repository = "https://github.com/virt-do/cloudlet"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.76.0"
 
 [dependencies]
+clap = { version = "4.5.1", features = ["derive", "env"] }
+clap-verbosity-flag = "2.2.0"
 epoll = "4.3.3"
 kvm-bindings = { version = "0.7.0", features = ["fam-wrappers"] }
 kvm-ioctls = "0.16.0"
 libc = "0.2.153"
 linux-loader = { version = "0.11.0", features = ["bzimage", "elf"] }
+log = "0.4.20"
 tracing = "0.1.40"
+tracing-subscriber = "0.3.18"
 vm-device = "0.1.0"
 vm-memory = { version = "0.14.0", features = ["backend-mmap"] }
 vm-superio = "0.7.0"
