+++ src/vmm/src/core/cpu/mpspec.rs
