+++ src/vmm/src/core/devices/mod.rs
